Time: 18 ms (62.72%), Space: 11.8 MB (59.57%) - LeetHub

+++ 462-minimum-moves-to-equal-array-elements-ii/462-minimum-moves-to-equal-array-elements-ii.cpp
@@ -5,7 +5,7 @@ public:
         ll n = nums.size(), steps = 0;
         nth_element(nums.begin(), nums.begin()+(n/2), nums.end()); 
         ll median = nums[n/2];
-        for(int i=0; i<n; i++){
+        for(ll i=0; i<n; i++){
             steps += abs(nums[i] - median);
         }
         return steps;
